# contracts/catalog/openapi.json
{
  "components": {
    "schemas": {
...
      "CanonicalError": {
        "description": "Error envelope. `field` + `reason` populated for param_invalid; both optional.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/CanonicalErrorCode"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "CanonicalError",
        "type": "object"
      },
      "CanonicalErrorCode": {
        "description": "Structured error codes for the canonical endpoint. HTTP status mirrors these\n(see /devices/{id}/canonical responses): 404 for the three 'not supported' codes,\n400 for param_invalid, 503 for device_unreachable, 500 for internal_error.",
        "enum": [
          "device_not_found",
          "capability_not_supported",
          "action_not_supported",
          "param_invalid",
          "device_unreachable",
          "internal_error",
          "no_active_scenario",
          "role_unbound",
          "no_group_members",
          "no_default_device",
          "fanout_not_allowed"
        ],
        "title": "CanonicalErrorCode",
        "type": "string"
      },
...
      "EvidenceEnvelope": {
        "description": "The bundle-shaped evidence — returned by ``GET /reports/evidence`` (redacted,\nno filing) and embedded in every filed bundle. Top-level keys are the contract.",
        "properties": {
          "bridge": {
            "additionalProperties": true,
            "description": "version, platform, catalog_version",
            "title": "Bridge",
            "type": "object"
          },
          "configs": {
            "additionalProperties": true,
            "description": "scoped device configs, redacted",
            "title": "Configs",
            "type": "object"
          },
          "context": {
            "additionalProperties": true,
            "description": "entity_id/room the report is anchored to + the scoped device set",
            "title": "Context",
            "type": "object"
          },
          "dispatch_ring": {
            "description": "last executed actions (B-2)",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Dispatch Ring",
            "type": "array"
          },
          "generated_at": {
            "description": "UTC ISO-8601 timestamp of collection",
            "title": "Generated At",
            "type": "string"
          },
          "logs": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "log filename -> base64(gzip(content))",
            "title": "Logs",
            "type": "object"
          },
          "mqtt_window": {
            "description": "recent broker traffic (B-2)",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Mqtt Window",
            "type": "array"
          },
          "scenarios": {
            "additionalProperties": true,
            "description": "active scenario + manual steps per room",
            "title": "Scenarios",
            "type": "object"
          },
          "state_diffs": {
            "additionalProperties": true,
            "description": "persisted-vs-live field diffs for the scoped devices",
            "title": "State Diffs",
            "type": "object"
          },
          "states": {
            "additionalProperties": true,
            "description": "live state of EVERY device (B-1)",
            "title": "States",
            "type": "object"
          },
          "system_config": {
            "additionalProperties": true,
            "description": "system.json, redacted",
            "title": "System Config",
            "type": "object"
          }
        },
        "required": [
          "generated_at",
          "bridge"
        ],
        "title": "EvidenceEnvelope",
        "type": "object"
      },
...
      "GroupMemberResult": {
        "description": "Per-member outcome of a room group action. `skipped` = the member's\nmatching capability lacks the requested action (reported, never an error);\n`no_op` = already at target; `failed` carries the member's error in `detail`.",
        "properties": {
          "capability": {
            "description": "The member's OWN capability the action ran against (a light switch's 'power', the hood's 'light').",
            "title": "Capability",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "device_id": {
            "title": "Device Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "executed",
              "no_op",
              "skipped",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "device_id",
          "capability",
          "status"
        ],
        "title": "GroupMemberResult",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReportRequest": {
        "properties": {
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Page context: route, entity_id — anchors the evidence scoping (B-1)",
            "title": "Context"
          },
          "free_text": {
            "description": "The user's problem description, verbatim",
            "minLength": 1,
            "title": "Free Text",
            "type": "string"
          },
          "ui_evidence": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Browser-side evidence (B-4): action log, console/API rings, SSE health, app context",
            "title": "Ui Evidence"
          }
        },
        "required": [
          "free_text"
        ],
        "title": "ReportRequest",
        "type": "object"
      },
      "ReportResponse": {
        "properties": {
          "report_id": {
            "title": "Report Id",
            "type": "string"
          },
          "spooled": {
            "description": "True = delivery failed, the report is spooled and will retry (B-7)",
            "title": "Spooled",
            "type": "boolean"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Ticket URL when filed immediately",
            "title": "Url"
          }
        },
        "required": [
          "success",
          "report_id",
          "spooled"
        ],
        "title": "ReportResponse",
        "type": "object"
      },
...
      "RoomCanonicalRequest": {
        "description": "Request to invoke a canonical action on a room GROUP — the third canonical\naddress form, for utterances that name a capability rather than a device.",
        "properties": {
          "action": {
            "description": "Action within each member's matching capability (e.g. 'on', 'off', 'open', 'close').",
            "title": "Action",
            "type": "string"
          },
          "group": {
            "description": "Semantic group name (e.g. 'light', 'cover') — a capability's group defaults to its domain name; profiles override (light_switch power → 'light').",
            "title": "Group",
            "type": "string"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          },
          "scope": {
            "default": "auto",
            "description": "auto = the room's configured default device for the group, else fan-out; all = force fan-out (the plural/«весь» signal); one = default device required (409 no_default_device if the room declares none).",
            "enum": [
              "auto",
              "all",
              "one"
            ],
            "title": "Scope",
            "type": "string"
          },
          "wait": {
            "default": true,
            "description": "Per-member echo-wait (same semantics as the device endpoint's `wait`).",
            "title": "Wait",
            "type": "boolean"
          }
        },
        "required": [
          "group",
          "action"
        ],
        "title": "RoomCanonicalRequest",
        "type": "object"
      },
      "RoomCanonicalResponse": {
        "description": "Response envelope for `POST /rooms/{room_id}/canonical`. 200 even with partial\nmember failures (the caller decides how to speak them); `success` = at least one\nmember executed or was already at target. Error envelopes (404/409) carry `error`\nand an empty `results`.",
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CanonicalError"
              },
              {
                "type": "null"
              }
            ]
          },
          "group": {
            "title": "Group",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/GroupMemberResult"
            },
            "title": "Results",
            "type": "array"
          },
          "room_id": {
            "title": "Room Id",
            "type": "string"
          },
          "scope_applied": {
            "anyOf": [
              {
                "enum": [
                  "default",
                  "fan_out"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Applied"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "room_id",
          "group",
          "action"
        ],
        "title": "RoomCanonicalResponse",
        "type": "object"
      },
      "RoomDefinitionResponse": {
        "description": "Response model for room definitions.",
        "properties": {
          "default_scenario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Scenario"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "devices": {
            "items": {
              "type": "string"
            },
            "title": "Devices",
            "type": "array"
          },
          "group_defaults": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Defaults"
          },
          "names": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Names",
            "type": "object"
          },
          "room_id": {
            "title": "Room Id",
            "type": "string"
          }
        },
        "required": [
          "room_id",
          "names",
          "description",
          "devices"
        ],
        "title": "RoomDefinitionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "A web service that manages MQTT devices with typed configurations",
    "title": "MQTT Web Service",
    "version": "0.6.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/reports": {
      "post": {
        "description": "File a problem report: collect evidence, package the envelope, deliver (or spool).",
        "operationId": "file_problem_report_reports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "File Problem Report",
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/evidence": {
      "get": {
        "description": "The bundle-shaped, redacted evidence — no ticket filed (B-11 read seam).",
        "operationId": "get_report_evidence_reports_evidence_get",
        "parameters": [
          {
            "description": "Anchor entity for B-1 scoping (optional)",
            "in": "query",
            "name": "entity_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Anchor entity for B-1 scoping (optional)",
              "title": "Entity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Report Evidence",
        "tags": [
          "reports"
        ]
      }
    },
...
    "/room/{room_id}": {
      "get": {
        "description": "Get a specific room definition.\n\nArgs:\n    room_id: The ID of the room to retrieve\n\nReturns:\n    RoomDefinitionResponse: The room definition\n\nRaises:\n    HTTPException: If room not found or service not initialized",
        "operationId": "get_room_room__room_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "title": "Room Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomDefinitionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Room",
        "tags": [
          "Rooms"
        ]
      }
    },
    "/rooms/{room_id}/canonical": {
      "post": {
        "description": "Room-scoped group actuation — the third canonical address form.\nFor utterances that name a capability, not a device\n(«включи свет», «закрой шторы»): the caller supplies room + group + action, the\nbridge owns membership (the `group` overlay on capability maps) and the\ndefault-vs-fan-out policy (`scope` + the room's `group_defaults`).\n\nMembers execute concurrently through the SAME per-device canonical dispatch the\ndevice endpoint uses (per-member no_op short-circuit, echo wait, `update_state`\nchokepoint — all unchanged). 200 even with partial member failures; the per-member\n`results` list is what makes the caller's confirmation honest.\n\nSpeakable errors:\n  - `404 no_group_members` — nothing in this room belongs to the group;\n  - `409 no_default_device` — scope=one but the room declares no default;\n  - `409 fanout_not_allowed` — a fan-out would be required but the group is not on\n    the allow-list (`light`, `cover`); consequential groups (`power`, …) need an\n    explicit device or scenario.",
        "operationId": "execute_room_canonical_action_rooms__room_id__canonical_post",
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "title": "Room Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCanonicalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomCanonicalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomCanonicalResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomCanonicalResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Room Canonical Action",
        "tags": [
          "Rooms"
        ]
      }
    },
...
  }
}